factor out upload logic
